[swarm_intrinics] Add manual task boundary annotation.
This commit relies on updates to SCC and related headers to achieve good
performance.  Please pull the latest changes to SCC and the Swarm
benchmarks repo and rebuild SCC before using this commit.

Each iteration of the hottest loop in pagerank traverse an edge, which
involves two things:
1.) Reading from a bunch of locations in memory find the src node ID,
    dest node ID, and the amount of the contribution from the src.
    These are all "read-only" data structures which aren't being updated
    frequently at the same time.
2.) Performing a read-modify-write on the the dest node's value.

While 1 accounts for more of the execution time and should be highly
parallelizable, 2 sometimes generates conflicts by accessing frequently
written cache lines.  During a conflict, the entire loop-iteration task
would be aborted.  Moreover, because the task needs to dynamically
obtain the destination node ID, which is not an induction variable and
is not known until after the task starts execution, the task cannot have
a spatial hint.  The manual task annotation has two benefits:
A.) By putting 1 and 2 in separate tasks, 2 can abort when needed to
    resolve conflicts without aborting 1.
B.) After 1 gets the dest node ID, it can spawn 2 and 2 can have a
    spatial hint, which prevents a lot of conflicts.

See Sections II.B and V.B of the T4 paper for background.
See comments in the "scc/hacks.h" header file for more information about
how to use annotations like this.